DS-3651: Fixes and refactoring
